clusters: search in nearest clusters too
As an improvement over the previous implementation, store each point's
n nearest clusters in it. This also allows searching in the nearest
clusters instead of only in the cluster that the point is participating.
This introduces one more hyperparameter we need to benchmark: the number
of nearest clusters to find and maintain for each point.

TODO: Add benchmarks.
TODO: Add comment describing the other implementation m+1

Signed-off-by: Plato Kiorpelidis <kioplato@gmail.com>